requirements.txt
# Core RAG dependencies
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-experimental>=0.0.50
langchain-community>=0.0.20

# Document processing
docling>=1.0.0
PyPDF2>=3.0.0
python-docx>=0.8.11

# Vector search and embeddings
faiss-cpu>=1.7.4
openai>=1.0.0

# BM25 for fusion retrieval
rank-bm25>=0.2.2

# Google Gemini
google-generativeai>=0.3.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
python-dotenv>=1.0.0

# Async processing
asyncio
concurrent.futures

# Logging and monitoring
loguru>=0.7.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Optional: for enhanced text processing
nltk>=3.8
spacy>=3.6.0

# Optional: for better tokenization
transformers>=4.30.0
tokenizers>=0.13.0
